requirements.txt
Django
django-ckeditor-5
django-filter
django-jazzmin
django-mptt
djangorestframework
drf-yasg
python-dotenv
Pillow
